+++ Estrapy/nsfw.py
@@ -62,7 +62,7 @@ class Nsfw:
 
         if generate:
             output = await self.base.produce(total=generate, route=route, type="link")
-            properties = PropertiesManager(urls=output)
+            properties = PropertiesManager(url=output)
 
         return properties
 
@@ -111,7 +111,7 @@ class Nsfw:
 
         if generate:
             output = await self.base.produce(total=generate, route=route, type="link")
-            properties = PropertiesManager(urls=output)
+            properties = PropertiesManager(url=output)
 
         return properties
 
@@ -160,6 +160,6 @@ class Nsfw:
 
         if generate:
             output = await self.base.produce(total=generate, route=route, type="link")
-            properties = PropertiesManager(urls=output)
+            properties = PropertiesManager(url=output)
 
         return properties